AB#983 Inherit affiliation from transitive parents
